requirements.txt
fastapi>=0.109.2
uvicorn>=0.27.1
python-dotenv>=1.0.1
yt-dlp>=2023.11.16
openai>=1.12.0
supabase>=2.3.4
pydantic>=2.7.0
requests>=2.31.0
